fixed circular import for debug and added a memory usage function
